修复文章编辑，添加报错
